Updated openstack/openstack Project: openstack-infra/project-config  8c0c931ab90219c724a1c771a769cd202da99ea6
Fix glance store experimental test

I'm not too familiar with this side of the code. I noticed a few other
tests have this line and the glance_store functional tests are missing it.

The error is that "The following LIBS_FROM_GIT were not installed
correct:  glance_store". From the latest check experimental on this patch:
https://review.openstack.org/#/c/302374/

Change-Id: I11a6d898ea3ee3b28c974cc6b99c9c9d87ef9e99