Updated openstack/openstack Project: openstack/trove  3161fe62023ee0ab0659537c945c10601edd24c2
PG test-helper create matching db for test user

Postgres scenario tests currently skip data operations because
the test helper fails to connect to the database.

There must be a database with the same name as the user in order
for the user to be able to login.
Once logged in the schema has to be referenced as 'public'.

Scenario tests run, with this fix applied.

* Also changed the dynamic configuration test property - the previous
  value was only in v9.4 and higher.

Change-Id: I4ba124196ff17c245b11fd13dd6bac9d2407c934