Updated openstack/openstack Project: openstack-infra/project-config  f2fb5cfe134563cba50d90d8d3f5f2a4989a28e2
Revert "Change kolla gate timeout to 90 minutes"

I'm not realy sure what the problem is with our gate timing out,
but 90 minutes doesn't fix the problem.  It just prolongs the delay
for gating when the gate fails in some mysterious way.

Thanks for working with us while we sort out our gating situation.

This reverts commit 01155c491b0b5e6abb27e2a06181ea4a59cc5eba.

Change-Id: I45ff5b59697d949e2c9ea8412442b4649486b2ea