Updated openstack/openstack Project: openstack-infra/nodepool  7c9c0516c54097159e1307fc5b5b48467d8faa8b
Write public IP into private when there's no private

devstack-gate multi-node depends on there always being a value in
private_v4 - even if that value is the public address. That's clearly
going to lead to confusion long term - but until we can rework it, this
should unbreak jobs that depend on that interface.

Change-Id: Icf52c9358cfe7c5103ce785027730e7c06c4429e