Updated openstack/openstack Project: openstack/cinder  2596004a542053bc19bb56b9a99f022368816871
Stop using tpool with oslo_db

oslo_db no longer supports the 'use_tpool' configuration option and
we see a warnings to that effect [1] whenever we run unit tests.

This commit modifies the cinder db api so that when in interacts
with oslo_db it no longer uses the db_concurrency.TpoolDbapiWrapper
and thereby avoids setting the 'use_tpool' configuration option.

[1] oslo_db/sqlalchemy/enginefacade.py:241: NotSupportedWarning:
    Configuration option(s) ['use_tpool'] not supported


Change-Id: I4e7fcc8fe86a1a22224c929e42bdf765e25131c8