Updated openstack/openstack Project: openstack/keystone  a957deccf46fe7048d103c075cf0a6d0d94113f8
Add logging to cli if keystone.conf is not found

If keystone.conf is not found at one of ./etc/keystone.conf,
~/.keystone/keystone.conf, ~/keystone.conf,
/etc/keystone/keystone.conf, or /etc/keystone.conf, the keystone-manage
command will use defaults configured in keystone.common.config[1] (or
elsewhere, e.g. keystone.common.sql.core[2]). If it does not find a
default value for a parameter it needs to use, it will error at that
point (for example, the samp_idp_metadata command errors when
idp_entity_id is not set). However, if all of the parameters it is
using have default values, which is the case for commands like db_sync
and ssl_setup, keystone-manage will silently proceed with those
defaults. This is not obvious to the user, who may have misplaced the
keystone.conf or is lacking permissions to read keystone.conf. This
patch adds a warning that will notify the user if it can't find the
config file, but otherwise proceeds as normal.

Why not fix this in oslo.config? The behavior of silently continuing if
config files aren't found is longstanding, so changing that behavior
would probably not be backwards-compatible. Moreover, other projects
might want to handle this differently or not handle it at all.

[1] http://git.openstack.org/cgit/openstack/keystone/tree/keystone/common/config.py
[2] http://git.openstack.org/cgit/openstack/keystone/tree/keystone/common/sql/core.py#n73

Closes-bug: #1273273

Change-Id: I276c671a0da78e3d1d2aa7336e55f65be41d8cca