Updated openstack/openstack Project: openstack/cinder  4978ce239e1ed9adf180d0cd8d83ca6295186f9a
Implement Cluster Pairing for SolidFire Driver

Add the ability to create remote Cluster pairs for
the SolidFire driver.  This is limited to replication_device
entries in the conf file for now, but can/will be expanded later.

For now, we simply check if there are replication_devices
specified in the config file, then querie the Cluster to see if
the pairing is already preent, if not, create the pairing and store
the info in an Object attribute.

This will allow us to just "swap" the active_cluster when a failover
is requested.

Change-Id: I8d7378ff5e96e35a11fa3e5b19c7905445141327