Updated openstack/openstack Project: openstack/heat  04b595b0ae3407e118df755d05cefc859e65ac30
Reinstate deprecation warnings

The commit 4188c40bebe1ab50e7b029584ab8a0fee8d5725e cleaned up a lot of
places where the warnings module was just being used as a non-standard way
to log a warning. However, the way the module was used in the rsrc_defn
module is correct and should not have been changed. By passing a
DeprecationWarning to warnings.warn() we not only make it easy to find
deprecated functions in the source, we also make it possible for developers
to test whether deprecated code is unused by forcing the exception to be
raised. There's also no point in translating these warnings, since they are
targeted at developers (mostly developers of third-party plugins) and not
operators per se. This patch reverts the change to that one file and its
tests.

Change-Id: I9cb1e739e9f173c8f70cb944e886cc3b49abc5f3