Updated openstack/openstack Project: openstack/zaqar  d31c96924e1851a8c3b72696424bf3161809c978
Merge tag '2.0.0'

This is a null-merge of the 2.0.0 release tag back into the master
branch so that the 2.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iae0d9bc98a6dad492dd0c25ccb0fb9888063e817