Updated openstack/openstack Project: openstack/cinder-specs  a9d9a8b873d9c00a85dbb2b1917a5d840c4d901d
Support backup import on another Storage database

Currently a volume backup can only be restored on the same Block
Storage service. This is because restoring a volume from a backup
requires metadata available on the remote database.

This proposal is to store backup metadata together
with the data, so a user will be able to restore a volume backup
on another Block Storage.

Change-Id: Ibafe96a3d1dcee8b6f28dac2f18814d3b7fcd9ef
Implements: blueprint support-backup-import-on-another-storage-database