Updated openstack/openstack Project: openstack/oslo.db  38167392f6aa1b9b084204895bb7f41723afb334
Add debtcollector to requirements

debtcollector is used by oslo.db at runtime, but is not currently
mentioned in requirements.txt explicitly (which effectively means
it's installed as a transitive dependency now).

This fixes tox -epip-missing-reqs run. A follow up change to
project-config will make sure we run this check in the gate.

Closes-Bug: #1571739

Change-Id: I564030830708da5c8a18f0c8a127ae74d54ead39