Updated openstack/openstack Project: openstack/keystone  018db7af3ca79cefcb8bbc01b6f1370b99c58734
Test list project hierarchy is correct for a large tree

Looking at the tests for get_projects_in_subtree_as_ids() I could not find
a test that checks if a large project tree is returned correctly.
In this test I ensure the order of the projects are returned correctly.

Change-Id: I569f70f6ec479f8cf140485a049fae0786996bbc