Updated openstack/openstack Project: openstack-infra/project-config  ea7fbca4d2c939bb296894bf5ebbcca2b97cbc39
Add release notes jobs for trove-dashboard CI

This job adds the CI jobs for Trove Dashboard. The change 3e0ad3e
already added the release notes jobs to the Trove Dashboard project's
tox.ini etc.,

I'd followowed the documentation at [1] which didn't mention that you
had to add the doc-publisher-site. I have therefure also submitted [2]
to update the documentation and pointed the example to this review as
it contains the three steps that are required to completely execute
this process.

[1] http://docs.openstack.org/project-team-guide/release-management.html#how-to-add-new-release-notes
[2] https://review.openstack.org/#/c/307161/

Change-Id: I6adb4934e82b934e1c8aa992c42e2377ef7a9732