Updated openstack/openstack Project: openstack-infra/project-config  e28edcab2446232d7b18ef7f64b8c808f5394551
add releasenotes and beaker jobs on puppet-ovn

We would like to start implementing functional testing on puppet-ovn and
this patch is a first step for having those tests, with Beaker.
Once we have basic beaker tests, we'll continue testing with
puppet-openstack-integration later.

Also add reno jobs to manage release notes.

Change-Id: Ib70a4cc724d4a17d0c1eab952bd872ac46f76b0a