Updated openstack/openstack Project: openstack/python-troveclient  1084741cd04d3f60bc6e5547c8d655117a5cf492
Allow use of backup name in trove create

'trove create' only accepts a backup id as a value
when restoring an instance from a backup. This change
enables the use of the backup name as well.

A few minor typos were fixed as well to make the output
more consistent.

Co-authored-by: Sushil Kumar <skm.net@gmail.com>
Co-authored-by: Peter Stachowski <peter@tesora.com>
Change-Id: I01f9ba96c8bdec5cfa4c4b477eb824b4456774d1
Closes-Bug: #1513636