Updated openstack/openstack Project: openstack/tempest  326bbb4884e28fdd9d33e22bfa882ecaf35a1d4b
"is_resource_deleted" fails to verify delete

Is resource deleted in images_client.py isn't adequate
for admin role because when checking for deleted images
the admin role returns a value whether the image is active or deleted.
This causes an endless loop if a call is waiting on a response of true.

Change-Id: I4e6937c992398cfc7db5391cfabb101aaefdb7e6
Closes-Bug: #1557088
Closes-Bug: #1564661
Closes-Bug: #1564614