Updated openstack/openstack Project: openstack/nova  79fd7d99b8a4c7c57a9a49bca4e94ab94e552312
Update servers.inc to be as accurate as api-site

This updates servers.inc which had a rough path during conversion. The
GET servers/details was lost entirely.

It updates a number of parameters to distinguish ones that are only in
the queries, as well as fixes references to short names that collided.

It updates the UPDATE call to only specify request parameters which
you can actually update.

It introduces the notation object.field for parameters that are
embedded in other objects. This is a new concept, but one way we can
clear up that the uuid in the parameter list is really networks.uuid.

It also deletes the old hand rolled servers.rst, which drops the
warnings count by nearly 100.

Part of bp:api-ref-in-rst

Change-Id: Ie7693a5888efa913118244a0837e6c0177965d48