Updated openstack/openstack Project: openstack-dev/devstack  06f1ddd88603fcadf429825fd84ff8b8975d2894
worlddump: request Guru Mediation reports for neutron agents

Those reports may be helpful when debugging neutron gate issues.

pgrep is backwards compatible with old Solaris tools, which means it
does not match with commands that are longer than 15 characters. To
avoid that for neutron agent names which are longer than that, we need
to pass -f argument to match against the full cmdline.

Also killall instead of kill + pgrep in a subshell.

Change-Id: I9b3801e927c0e80443ed76e38cd8e3618e888e49