Updated openstack/openstack Project: openstack-dev/pbr  a05b55b3e36753c9a341eb39381b702c45a95412
Fix wsgiref script use with oslo.config

The script generated for wsgi-script entrypoint adds a argument for
listening on a specific port using argparse. Unfortunately it's not
compatible with oslo.config handling of config options, as used by
keystone for example. This fixes the situation by only parsing known
options.

Change-Id: I37f82e8d78a4288323854282da300c123561218a
Closes-Bug: #1501756