Updated openstack/openstack Project: openstack/neutron  d54e03d233c1cd160cbcf5bbdcfd289b5c1e93d1
Support interface drivers that don't support mtu parameter for plug_new

The method signature before Mitaka did not have the mtu= parameter. We
should continue supporting the old signature, since it can be used in
out of tree interface drivers. The class is part of public neutron API,
so we should make an effort to not break out of tree code.

Change-Id: I8e0c07c76fd0b4c55b66c20ebe29cdb7c07d6f27
Closes-Bug: #1570392