Updated openstack/openstack Project: openstack/cinder  6b2201421725d87a7c2faf3c8e8e85f6a3595975
Add failed_over detection to SolidFire driver

This completes the Cheesecake implementation of
replication for the SolidFire driver.

This final step adds the detection of backend_id
on init and sets the active cluster appropriately
using the config entries for "replication_device"
and the passed in backend_id that's sent in on
init based on the entry in the cinder db.

Change-Id: Iee5684c6f7a2199d1f0211fb3c22f6c1593bc604