Updated openstack/openstack Project: openstack/swift  d6471c70196a6f8b41db74e2c48d121ceccfb563
Stop staticweb revealing container existence to unauth'd requests

When a container has `X-Container-Meta-Web-Listings: false` then
staticweb will return a 404 in response to a GET or HEAD on the
container, regardless of whether the request is auth'd. That provides
a way to probe for container existence. It should return a 401 if the
request is not auth'd.

This patch adds a call to swift.authorize before returning the 404.

Closes-Bug: 1506116
Change-Id: I382323b49dc8f6d67bf4494db7084a860a10db59