Updated openstack/openstack Project: openstack/ceilometermiddleware  cacc8521a5c318284d5bd14f5a4bfdf49ff94640
use path as resource id when missing RESELLER prefix

we don't have an id if RESELLER prefix is not in path. this causes
an id to be autogenerated making each request a completely different
resource id.

this patch changes it so we just default to path as our id if one
cannot be computed.

Change-Id: I5775406a84b5f2ad0f95f366b7deca210d2f8e87
Closes-Bug: #1566940