Updated openstack/openstack Project: openstack/cinder  0f9cf6c20c159efdedb8d0438af20b7369d0bdae
Pylint: refactor ignored message lists

The way these are organized makes it rather hard to
associate comments and their messages, and makes it
more work to modify this list in a meaningful way (or
re-use it for a project like os-brick).

Move things around a bit to make this easier to manage.

Change-Id: Ieebca9f97f6b15832693a37fa03f97c1074d64e8