Updated openstack/openstack Project: openstack-infra/storyboard-webclient  f7c8ae07096120605b53fcc09a0d972bdb1eab6b
Don't allow users to interact with cards in automatic worklists

This commit stops users from being able to drag cards in/out of
automatic worklists, and also hides the due date information on
the card modal if the card is in an automatic worklist.

Also, this makes automatic worklists be visually different from
manual worklists in boards.

Change-Id: I78da3e0dea2e420b9bcfec856f92afd1558287bf