Updated openstack/openstack Project: openstack/neutron  8f33cbeb4264d059ce4daa933c5381ab54e34439
Revert "Deprecate 'ovs_use_veth' and 'veth_mtu' options"

The reason for reverting this is that when we tested this OVS
port is show as DOWN and it can actually send traffic.

An explanation is below:

The DHCP port has "tap" in the name, but doesn't appear to be
a tap device. It appears to be an OVS internal device. That device
was then put into a different network namespace, which isn't
visible to the root namespace in which ovs-vswitchd is running.
Using internal devices in this manor is kind of a hack since
ovs-vswitchd can't actually see the device, but it can still
send and receive traffic on it.

This reverts commit 850b4c025fa5652bfa474c78e7775d74fdb93df6.

Closes-bug: #1571553
Change-Id: I82d53d62bfd42cb17e3ce7f739369be3bbf44c02