Updated openstack/openstack Project: openstack-infra/project-config  5f33c590cfe12a318e96b9cc8bd294fda8b95f1f
Move OpenStack-Ansible CentOS7 jobs to experimental queue

As the work to implement CentOS7 support in OpenStack-Ansible is
low priority work which is being done over several cycles, the
jobs to implement this are being moved to the experimental queue.

This will reduce the OpenStack-Infra resources wasted and only
implement the jobs on-demand.

Change-Id: I52e044a6d47691cca52e6c76f6b6268c01003434