Updated openstack/openstack Project: openstack/nova  d2d5f10175eb13f4775f34657530a9514608d9da
Fix database poison warnings, part 3

The following warning appears in the unit test logs a number of times.
This patch fixes the warnings in ComputeAPIAPICellUnitTestCase.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: Idcd733df2aafe809cb89f2a1db5df6ad227dd39c
Related-Bug: #1568414