Updated openstack/openstack Project: openstack/nova  3316fa85668ee96f4db1bbff0752de25ce4744af
Remove deprecated ability to load scheduler_host_manager from path

The ability to load the scheduler_host_manager from a random
classpath was deprecated in Mitaka. This removes that ability
so that only the entrypoints in Nova's setup.cfg can be used
to load the scheduler host manager.

The scheduler_host_manager config option is updated to only
allow the choices that are currently in the setup.cfg.

And a release note is added for the upgrade impact of dropping
the deprecated classpath loading ability.

Change-Id: I886c74afac02d100d472b78522e8295748276e4d
Related-Bug: #1403020