Updated openstack/openstack Project: openstack/requirements  a166e15e5f61978b20dd6c27501d328b52238bef
Add xmltodict to global-requirements

Introduction:

xmltodict (by Martin Blech) is a Python module that facilitates
seamless xml-dict conversions as in this "spec" [1] [2]

In Trove we often need to read, parse, update and write configuration files
in many different formats. The configuration files are deserialized into,
handled as and serialized from native Python dicts.
Recently a need has arisen to be able to easily update an
XML configuration file. [3] This is exactly a task for 'xmltodict'.

Python support:

    The library currently supports Python 2.x to 3.5.

Community and development:

    As of writing this the latest update was on '23/02/2016'.
    The project has a public Git repo and is being activelly developed. [3]

License:

    It's being distributed under the MIT license.

No change to upper-constraints.txt:

    No change is being made to upper-constraints.txt because it
    already has a line pinning xmltodict to 0.10.1 which is the
    current version on pypi.

[1] XML to JSON: http://www.xml.com/pub/a/2006/05/31/
                 converting-between-xml-and-json.html
[2] xmltodict PyPi: https://pypi.python.org/pypi/xmltodict
[3] Related Trove change: Ia57587bf557771ea6d7f00e2fe9d674789b98559
[4] xmltodict GitHub: https://github.com/martinblech/xmltodict

Change-Id: I4a088b7a541c35a6c32d6985727bad65ff491815