Updated openstack/openstack Project: openstack-infra/project-config  a745444754d66caf2c8cd85e695237c7f09b54ac
Run slave-cleanup on all proposal jobs

slave-cleanup cleans up more than 100 MB for each job it runs - and it
cleans the same content that we normally clean when starting the job.
so, this will save some space on the proposal slave.

Add the slave-cleanup run to all jobs running on proposal node.

Change-Id: Ide862b0d1d066adbf992de7f542be2d4d20cf9b1