Updated openstack/openstack Project: openstack/keystone  ef07ad60b4c91271ad9438fb4a043ef36965e862
Use messaging notifications transport instead of default

The usage of oslo_messaging.get_transport is not meant for
notifications; And oslo_messaging.get_notification_transport is
favored for those means. So this change introduces the usage of that
function.

If the settings for the notifications are not set with the
configuration that's under the oslo_messaging_notifications group,
this will fall back to the old settings which are under the DEFAULT
group; just like this used to work.

Change-Id: Id1f89c89fa23d19e9229af7e80df179b6e17a635