Updated openstack/openstack Project: openstack/cinder-specs  12dd4ef1053a0cd26c52f8e256401dae0ff1ab83
User facing error Summary Messages

The purpose of this spec is to define a mechanism in Cinder that will
allow a user to see error messages for asynchronous operations.

This is more specific to failed operations and cause of their failure.

Spec started in Newton

APIImpact New /messages resource

DocImpact

Co-Authored-By: Alex Meade <mr.alex.meade@gmail.com>

Implements : blueprint summarymessage

Change-Id: If7f8b7f277863a6dbf9686dadab145590ec9567d