Updated openstack/openstack Project: openstack/ironic  f049cfd1888842e545ee28f4c18c92bc7ea3b763
Update the inspection documentation

This change moves inspection documentation to its own file
(similar to RAID and cleaning).

It also adds a list of capabilities we *might* discover during inspection.
It's partly based on the iLO spec [1], partly - on the in-band capabilities
discovery RFE [2]. The vendor-specific bits are missing on purpose.

The goal here is to have a common subset of capabilities to be discovered
by all (or the majority) of the drivers (including in-band inspection).

[1] http://specs.openstack.org/openstack/ironic-specs/specs/kilo-implemented/ilo-properties-capabilities-discovery.html#proposed-change
[2] https://bugs.launchpad.net/ironic-python-agent/+bug/1571580

Related-Bug: #1571580
Change-Id: If5d466a2c331674abb30d30dc39871f25a84de88