Updated openstack/openstack Project: openstack-infra/tripleo-ci  7d4fa5744d5d72c31c728204ad54258e4fad73ed
Conditionally request a test env from geard

Only request a test environment from geard if we don't
have one defined. This will help allow the toci script
to be run without the need for a geard managing the
environments. Also if the testenv wasn't requested from
geard leave it running to allow debugging.

Change-Id: Id2b03456be86e7f2900388f40668d79861858bf6