Updated openstack/openstack Project: openstack/nova  fad934a5ff471cff1b69642d2fa8838d0b99a3f6
Fix database poison warnings, part 2

The following warning appears in the unit test logs a number of times.
This patch fixes some of those warnings by stubbing at the object layer
rather than the database layer.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I29dca2fd07c7b106a51fb1608caa87b1a5548863
Related-Bug: #1568414