Updated openstack/openstack Project: openstack/nova  327ffd696e84b2d713342a1540e0a232cf96d59a
Fix test for empty policy rules

The test that was supposed to ensure the sample policy file does not
have rules defined as "" was not working as intended. It was taking the
parsed list of rules and checking for emptiness however empty rules are
converted to "@" rules when parsed. The test has been updated to look at
the actual policy definitions, not the parsed output.

Change-Id: I734b6929769bf394721c3c6f612e995d1718b647