Updated openstack/openstack Project: openstack-infra/shade  68f5bf94dae7e4f330fca7267583f25e41e708ea
Have delete_server use the timed server list cache

Waiting for server deletion is just as costly as waiting for creation.
Make sure that delete honors the 5-second cache if it's in use by using
get_server() instead of get_server_by_id().

Using the get_server() call identified a bug in the volume cache
invalidation that is done in the code immediately after the call, so we
fix that here too, and add a new cache_enabled attribute to allow us
to avoid an unnecessary API call.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>

Change-Id: I70ccfffe6cbb1f46049b3525bba35c612a572ef0