Updated openstack/openstack Project: openstack-infra/shade  28335c6c96acb7f465c78cd6d36bd34234f87000
Retry floating ip deletion before deleting server

If we're waiting for a server to be deleted, and we've asked shade to
delete the ips associated with the server, it stands to reason that the
action should not be considered complete until both things have been
accomplished.

Unfortunately, what's going on here is that deleting a floating ip is a
synchronous event that just sometimes doesn't work. It is, however, a
quick and not very costly event, and we've been doing it before the
delete call regardless of user choice on "wait" anyway. So just add a
configurable retry.

Change-Id: Id7b81417a59c00d3daeb49bd11df0e9fe7dc3b42