Updated openstack/openstack Project: openstack/neutron-vpnaas  9fc64611e53ae6ab514ed044f9e35c1b4ff6b7b7
Openswan/Libreswan: Check config changes before restart

Currently, when neutron-vpn-agent restart, all the pluto process in
router ns will be restarted too. But actually this is not required
and will impact the vpn traffic. In this patch, we keep a backup for
ipsec.conf and ipsec.secrets, and then compare the configurations
when restart, if no config changes. Restart can be skipped.

Note: this change is DocImpact

Change-Id: I5a7fae909cb56721bd7e4d42999356c7f7464358
Closes-Bug: #1571455