Updated openstack/openstack Project: openstack/heat  0879b4155483a420ef36b31f20ec46151dbb0744
Break reference cycle between Environment and ResourceRegistry

The commit 08431c7c0601f64e6f0477dd502bf912eba8529b added a reference to
the Environment object from the ResourceRegistry object, which is itself
an attribute of the environment. This causes a reference cycle that
means that the environment and everything referenced by it (including
the potentially very large files dict) will only be deallocated when it
is garbage collected since the reference counts will never hit zero.
This has probably been contributing substantially to memory
fragmentation.

Change-Id: Ib251b5f5ffc07fe1a06f5e44124024831d4ba1b2
Partial-Bug: #1570974