Updated openstack/openstack Project: openstack/tempest  16863a1b4b16b53c3a73813d5dc6c3122f2d8cbc
Add pep8 check to use data_utils.rand_uuid()

There is a common method for generating random uuid value.
So it is nice to avoid pointing it out on human review to reduce
reviewing workload.
In addition, this patch removes str() around the method calls
because str() is called in the method.

Change-Id: Ic8caea352d27c479fdb765a699d277b59f1c5bb4