Updated openstack/openstack Project: openstack-infra/project-config  08001cc2b530ea30d9319e9c41d1145eb51d02b0
New project devstack-plugin-tar-installer

The main reason to create this new project is a request from the
community in the context of this devstack patch:
https://review.openstack.org/#/c/108714/

The objective is to enable devstack to install software
(QEMU and libvirt for the time being) from tar releases. The
project devstack-plugin-tar-installer will host the devstack plugin
which implements this feature.

Change-Id: Iddaf12ba79aa70b6045826568319a163e75347b6
Signed-off-by: Michele Paolino <m.paolino@virtualopensystems.com>