Updated openstack/openstack Project: openstack-infra/tripleo-ci  4a1b53bebd384a9f28be7cdd39fd5211ba158e61
Update deprecated commands in tripleo.sh (Old heat * calls)

Update the references of 'heat *' calls.

In this patch, calls to 'heat *' are replaced using the OpenStack CLI.
This change was applied to:
    --overcloud-delete option
    cleanup_pingtest method

To support backwards compatibility with Liberty a bash snippet was
used to redefine the openstack command, and call the previous 'heat *'
interface when apply.

Change-Id: Ie062f55e29a1b0cdaf6249dbb08815c31642b60c