Updated openstack/openstack Project: openstack/horizon  ca8d61cf55d05823909e408c014bde8f2108be8d
Net topology: Show console link only when useful

Right now we are showing the "open console" link on the
network topology screen for instances on each instance,
even if they are on a status that we know its not
going to allow the console to connect.

This patch makes it so the "open console" link will
only appear if the instance object has console data.

It also changes the order of getting the instance
data in the django part as to make less calls if we
dont need the console link.

Provides a list with known statuses that wont connect
to the console used to get the console data or not.

Change-Id: I536fee5186ac933b92a7dc01a2a5b82a6db0ae4c
Closes-Bug: #1543316