Updated openstack/openstack Project: openstack-infra/project-config  c8a3b9cc1c73a02c3ba149a8ff3c97b6f2e5ade2
Add new virtualbmc project under ironic's umbrella

VirtualBMC is a utility that allows virtual/fake BMCs to be created
handling the IPMI commands and converting it to calls to the
python-libvirt library. The project was created to allow drivers using
the IPMI protocol to be tested in the Ironic gate.

Related-Bug: #1570301
Needed-By: I24349ce1b5c0332157f1bf8bd0613338a15495e4
Change-Id: Idf02cb28316be1846bb5fa1988c3d47ce6f9c54c