Updated openstack/openstack Project: openstack/heat  5672c4649e7dab23cbbc57903d3a2b514a61a5bb
Use messaging notifications transport instead of default

The usage of oslo_messaging.get_transport is not meant for
notifications; And oslo_messaging.get_notification_transport is
favored for those means. So this change introduces the usage of that
function.

If the settings for the notifications are not set with the
configuration that's under the oslo_messaging_notifications group,
this will fall back to the old settings which are under the DEFAULT
group; just like this used to work.

Change-Id: I52d7a72dadd10ade1f5335f14c58391271e209c2