Updated openstack/openstack Project: openstack-infra/tripleo-ci  6e76d225954a45d6316c97150cde036ea87c65ed
Make the delorean-ci repository available when build overcloud images

tripleo.sh defaults DIB_YUM_REPO_CONF to the delorean-dep, delorean
and delorean-current in CI we have a 4th repostiory containing the
packages that got built during this CI run. Pass in DIB_YUM_REPO_CONF
to override the default so that the overcloud image build installs
from the delorean-ci repository if it needs to.

Change-Id: Idb2326f8560f49ab5504a545aef35c984a439cc2
Closes-Bug: #1560257