Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  feb8e465d6e35b2b78270adf482931f0804a4223
Add support for disable-strict-forbidden-file-verification

Enabling this option will allow an event to trigger a build if the event
contains BOTH one or more wanted file paths AND one or more forbidden
file paths.  In other words, with this option, the build will not get
triggered if the change contains only forbidden files, otherwise it will
get triggered.

Support for this was added in Plugin Version 2.16.0.

Change-Id: I860b076c2393909012c3ac9821559c9e8c9f7e28