Updated openstack/openstack Project: openstack/trove  c3169966ba570583cc60c0fa3fc853587e9a069d
Mysql replicas need to set binlog_format

For GTID replication we set the replica source (i.e. master)
to have a binlog_format of MIXED. We aren't setting this on
the replicas so they default to STATEMENT. This causes a problem
with certain "non-deterministic" functions (e.g. RAND()). This
changes replica config templates for Mysql and Percona to use
MIXED mode.

A new scenario test was introduced to validate that the
replicas have this set.

MariaDB doesn't appear to have this issue so it wasn't changed.

Scenario tests for Mysql, Percona and MariaDB were run to
validate this change.

Change-Id: I936cd9bc53a812af19653e9b5b472103fab2b6c1
Closes-bug: 1563541